removed errors

+++ src/main/java/com/billioncart/service/serviceImpl/AddressServiceImpl.java
@@ -1,13 +1,17 @@
 package com.billioncart.service.serviceImpl;
 
+import java.util.List;
+import java.util.stream.Collectors;
+
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
+import com.billioncart.exception.AddressNotFoundException;
 import com.billioncart.mapper.AddressRequestMapper;
 import com.billioncart.mapper.AddressResponseMapper;
 import com.billioncart.model.Account;
 import com.billioncart.model.Address;
-import com.billioncart.model.User;
 import com.billioncart.payload.AddressRequest;
 import com.billioncart.payload.AddressResponse;
 import com.billioncart.repository.AccountRepository;
@@ -39,6 +43,7 @@ public class AddressServiceImpl implements AddressService{
         return AddressResponseMapper.INSTANCE.toPayload(createdAddress);
 	}
 	
+	@Transactional
 	public void removeAddress(Long addressId) {
 		String username = UserDetailsUtils.getAuthenticatedUsername();
 		
@@ -46,6 +51,39 @@ public class AddressServiceImpl implements AddressService{
         Account existingAccount = accountRepository.findByUsername(username)
         		.orElseThrow(() -> new UsernameNotFoundException("Account not found"));
         
-        addressRepository.removeByUserAndAddressId(existingAccount.getUser(), addressId);
+        Address existingAddress = addressRepository.findById(addressId).orElseThrow(() -> new AddressNotFoundException("Address not found"));
+        
+        if(existingAddress.getUser().getUserId() == existingAccount.getUser().getUserId()) {
+        	addressRepository.removeByUserAndAddressId(existingAccount.getUser(), addressId);        	
+        }else {
+        	throw new AddressNotFoundException("Address not found");
+        }
+	}
+	
+	@Override
+	public AddressResponse updateAddress(Long AddressId, AddressRequest request) {
+		Address existingAddress = addressRepository.findById(AddressId).orElseThrow(() -> new AddressNotFoundException("Address not found"));
+		
+		Address address = AddressRequestMapper.INSTANCE.toEntity(request);
+		address.setAddressId(existingAddress.getAddressId());
+		address.setUser(existingAddress.getUser());
+		
+		Address updatedAddress = addressRepository.save(address);
+		return AddressResponseMapper.INSTANCE.toPayload(updatedAddress);
+	}
+	
+	@Transactional
+	@Override
+	public List<AddressResponse> getAddressesByUser(){
+		String username = UserDetailsUtils.getAuthenticatedUsername();
+		
+		Account existingAccount = accountRepository.findByUsername(username).orElseThrow(() -> new UsernameNotFoundException("Account not found"));
+		
+		List<Address> addresses = addressRepository.findAllByUser(existingAccount.getUser());
+		
+		return addresses.stream().map(addr ->{
+			AddressResponse response = AddressResponseMapper.INSTANCE.toPayload(addr);
+			return response;
+		}).collect(Collectors.toList());
 	}
 }
